adding TDD

+++ models/Item.js
@@ -1,13 +1,13 @@
-const {Sequelize, DataTypes, Model} = require('sequelize')
-const {sequelize} = require('../db')
+const { Sequelize, DataTypes, Model } = require('sequelize')
+const { sequelize } = require('../db')
 
 
-class Item extends Model { }
+class Item extends Model {}
 
 Item.init({
     name: DataTypes.STRING,
     description: DataTypes.STRING,
-    price: DataTypes.float(5, 2),
+    price: DataTypes.DOUBLE(5, 2),
     image: DataTypes.STRING
 }, {
     sequelize,
